+++ audionimbus/src/effect/ambisonics/panning.rs
@@ -1,6 +1,4 @@
-use super::super::AudioEffectState;
-use super::super::EffectError;
-use super::super::SpeakerLayout;
+use super::super::{AudioEffectState, EffectError, SpeakerLayout};
 use crate::audio_buffer::{AudioBuffer, Sample};
 use crate::audio_settings::AudioSettings;
 use crate::context::Context;