+++ audionimbus/src/effect/virtual_surround.rs
@@ -1,6 +1,5 @@
 use super::audio_effect_state::AudioEffectState;
-use super::EffectError;
-use super::SpeakerLayout;
+use super::{EffectError, SpeakerLayout};
 use crate::audio_buffer::{AudioBuffer, Sample};
 use crate::audio_settings::AudioSettings;
 use crate::context::Context;